Create functions that installs dependencies, also create error logger.
These functions check for and install the modules necessary to run the application. It will attempt to import the modules and will attempt to install them using pip if this fails. If the pip installation fails, the function will attempt to install pip and then try to install the modules again. If this doesn't work, it will dump an error to error.log and exit the program.

Speaking of, I created a simple function using the logging module that dumps errors to a file named 'error.log'